% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_burial_model.R
\name{carbon_burial_model}
\alias{carbon_burial_model}
\title{Estimate carbon burial in lakes}
\usage{
carbon_burial_model(input_parms, method_list, return_all = F)
}
\arguments{
\item{input_parms}{list;}

\item{method_list}{list;}

\item{return_all}{boolean;}
}
\description{
Function to calculate carbon burial in a lake
}
